# frontend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProfitLens",
    "description": "Deterministic-first AI business analytics for D2C brands.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/analytics/v2/scenario/run": {
      "post": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Scenario Run",
        "operationId": "scenario_run_analytics_v2_scenario_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioV2Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/v2/investigations/{month}": {
      "get": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Investigations",
        "operationId": "investigations_analytics_v2_investigations__month__get",
        "parameters": [
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/v2/alerts/{month}": {
      "get": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Alerts",
        "operationId": "alerts_analytics_v2_alerts__month__get",
        "parameters": [
          {
            "name": "month",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/v2/alerts/evaluate": {
      "post": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Custom Alert Evaluation",
        "operationId": "custom_alert_evaluation_analytics_v2_alerts_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomAlertEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/v2/exports": {
      "post": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Create Export",
        "operationId": "create_export_analytics_v2_exports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/v2/saved-views": {
      "post": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Saved View Create",
        "operationId": "saved_view_create_analytics_v2_saved_views_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSavedViewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Saved View List",
        "operationId": "saved_view_list_analytics_v2_saved_views_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page"
            }
          },
          {
            "name": "created_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created By"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/v2/saved-views/{saved_view_id}": {
      "get": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Saved View Get",
        "operationId": "saved_view_get_analytics_v2_saved_views__saved_view_id__get",
        "parameters": [
          {
            "name": "saved_view_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Saved View Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Saved View Delete",
        "operationId": "saved_view_delete_analytics_v2_saved_views__saved_view_id__delete",
        "parameters": [
          {
            "name": "saved_view_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Saved View Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/v2/workspaces": {
      "post": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Workspace Create",
        "operationId": "workspace_create_analytics_v2_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/v2/workspaces/{workspace_id}": {
      "get": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Workspace Get",
        "operationId": "workspace_get_analytics_v2_workspaces__workspace_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/v2/users/{user_id}/workspaces": {
      "get": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "User Workspaces",
        "operationId": "user_workspaces_analytics_v2_users__user_id__workspaces_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/v2/workspaces/{workspace_id}/members": {
      "post": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Workspace Member Add",
        "operationId": "workspace_member_add_analytics_v2_workspaces__workspace_id__members_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWorkspaceMemberRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analytics/v2/brands": {
      "post": {
        "tags": [
          "ProfitLens Platform V2"
        ],
        "summary": "Brand Create",
        "operationId": "brand_create_analytics_v2_brands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomAlertEvaluationRequest": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/AlertRule"
            },
            "type": "array",
            "title": "Rules"
          }
        },
        "type": "object",
        "required": [
          "month",
          "rules"
        ],
        "title": "CustomAlertEvaluationRequest"
      },
...
      "ExportBuildRequest": {
        "properties": {
          "export": {
            "$ref": "#/components/schemas/ExportRequest"
          },
          "data": {
            "title": "Data"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "export",
          "data"
        ],
        "title": "ExportBuildRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScenarioV2Request": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "changes": {
            "$ref": "#/components/schemas/ScenarioChanges"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "month",
          "changes"
        ],
        "title": "ScenarioV2Request"
      },
...
    }
  }
}